symfony3

+++ src/Kunstmaan/NodeSearchBundle/Configuration/NodePagesConfiguration.php
@@ -7,6 +7,9 @@ use Elastica\Index;
 use Elastica\Type\Mapping;
 use Kunstmaan\AdminBundle\Helper\DomainConfigurationInterface;
 use Kunstmaan\AdminBundle\Helper\Security\Acl\Permission\MaskBuilder;
+use Kunstmaan\NodeBundle\Entity\HasNodeInterface;
+use Kunstmaan\NodeBundle\Entity\Node;
+use Kunstmaan\NodeBundle\Entity\NodeTranslation;
 use Kunstmaan\NodeBundle\Entity\NodeVersion;
 use Kunstmaan\NodeBundle\Entity\PageInterface;
 use Kunstmaan\NodeBundle\Helper\RenderContext;
@@ -14,9 +17,6 @@ use Kunstmaan\NodeSearchBundle\Event\IndexNodeEvent;
 use Kunstmaan\NodeSearchBundle\Helper\IndexablePagePartsService;
 use Kunstmaan\NodeSearchBundle\Helper\SearchBoostInterface;
 use Kunstmaan\NodeSearchBundle\Helper\SearchViewTemplateInterface;
-use Kunstmaan\NodeBundle\Entity\HasNodeInterface;
-use Kunstmaan\NodeBundle\Entity\Node;
-use Kunstmaan\NodeBundle\Entity\NodeTranslation;
 use Kunstmaan\PagePartBundle\Helper\HasPagePartsInterface;
 use Kunstmaan\SearchBundle\Configuration\SearchConfigurationInterface;
 use Kunstmaan\SearchBundle\Provider\SearchProviderInterface;
@@ -142,6 +142,9 @@ class NodePagesConfiguration implements SearchConfigurationInterface
 
             //build new index
             $index = $this->searchProvider->createIndex($this->indexName . '_' . $locale);
+            if ($index->exists()) {
+                continue;
+            }
 
             //create index with analysis
             $this->setAnalysis($index, $localeAnalysis->setupLanguage($language));
@@ -289,6 +292,12 @@ class NodePagesConfiguration implements SearchConfigurationInterface
     public function deleteIndex()
     {
         foreach ($this->locales as $locale) {
+            /** @var Index $index */
+            $index = $this->searchProvider->getIndex($this->indexName . '_' . $locale);
+            if (!$index->exists()) {
+                continue;
+            }
+
             $this->searchProvider->deleteIndex($this->indexName . '_' . $locale);
         }
     }
@@ -512,7 +521,12 @@ class NodePagesConfiguration implements SearchConfigurationInterface
     protected function enterRequestScope($lang)
     {
         $requestStack = $this->container->get('request_stack');
-        if (!$requestStack->getCurrentRequest()) {
+        // If there already is a request, get the locale from it.
+        if ($requestStack->getCurrentRequest()) {
+            $locale = $requestStack->getCurrentRequest()->getLocale();
+        }
+        // If we don't have a request or the current request locale is different from the node langauge
+        if (!$requestStack->getCurrentRequest() || ($locale && $locale !== $lang)) {
             $request = new Request();
             $request->setLocale($lang);
